feat(supplier-invoices): a line's account is a category chip with the Transaktioner picker
The account cell shows the account as the Transaktioner category chip does
(hue dot, name, number) and opens the same anchored picker: choose a
template and the line books to its debit account, or search an account
directly. The combobox-in-the-cell is gone. AccountChip moves to
components/ui so both pages share it.
